+++ sdk/resourcemanager/devcenter/armdevcenter/responses.go
@@ -100,6 +100,25 @@ type CheckScopedNameAvailabilityClientExecuteResponse struct {
 	CheckNameAvailabilityResponse
 }
 
+// CustomizationTasksClientGetErrorDetailsResponse contains the response from method CustomizationTasksClient.GetErrorDetails.
+type CustomizationTasksClientGetErrorDetailsResponse struct {
+	// List of validator error details. Populated when changes are made to the resource or its dependent resources that impact
+	// the validity of the Catalog resource.
+	CatalogResourceValidationErrorDetails
+}
+
+// CustomizationTasksClientGetResponse contains the response from method CustomizationTasksClient.Get.
+type CustomizationTasksClientGetResponse struct {
+	// Represents a Task to be used in customizing a Dev Box.
+	CustomizationTask
+}
+
+// CustomizationTasksClientListByCatalogResponse contains the response from method CustomizationTasksClient.NewListByCatalogPager.
+type CustomizationTasksClientListByCatalogResponse struct {
+	// Results of the Task list operation.
+	CustomizationTaskListResult
+}
+
 // DevBoxDefinitionsClientCreateOrUpdateResponse contains the response from method DevBoxDefinitionsClient.BeginCreateOrUpdate.
 type DevBoxDefinitionsClientCreateOrUpdateResponse struct {
 	// Represents a definition for a Developer Machine.
@@ -259,6 +278,12 @@ type GalleriesClientListByDevCenterResponse struct {
 	GalleryListResult
 }
 
+// ImageVersionsClientGetByProjectResponse contains the response from method ImageVersionsClient.GetByProject.
+type ImageVersionsClientGetByProjectResponse struct {
+	// Represents an image version.
+	ImageVersion
+}
+
 // ImageVersionsClientGetResponse contains the response from method ImageVersionsClient.Get.
 type ImageVersionsClientGetResponse struct {
 	// Represents an image version.
@@ -269,6 +294,18 @@ type ImageVersionsClientGetResponse struct {
 type ImageVersionsClientListByImageResponse struct {
 	// Results of the image version list operation.
 	ImageVersionListResult
+}
+
+// ImageVersionsClientListByProjectResponse contains the response from method ImageVersionsClient.NewListByProjectPager.
+type ImageVersionsClientListByProjectResponse struct {
+	// Results of the image version list operation.
+	ImageVersionListResult
+}
+
+// ImagesClientGetByProjectResponse contains the response from method ImagesClient.GetByProject.
+type ImagesClientGetByProjectResponse struct {
+	// Represents an image.
+	Image
 }
 
 // ImagesClientGetResponse contains the response from method ImagesClient.Get.
@@ -285,6 +322,12 @@ type ImagesClientListByDevCenterResponse struct {
 
 // ImagesClientListByGalleryResponse contains the response from method ImagesClient.NewListByGalleryPager.
 type ImagesClientListByGalleryResponse struct {
+	// Results of the image list operation.
+	ImageListResult
+}
+
+// ImagesClientListByProjectResponse contains the response from method ImagesClient.NewListByProjectPager.
+type ImagesClientListByProjectResponse struct {
 	// Results of the image list operation.
 	ImageListResult
 }
@@ -415,6 +458,53 @@ type ProjectCatalogEnvironmentDefinitionsClientGetErrorDetailsResponse struct {
 	CatalogResourceValidationErrorDetails
 }
 
+// ProjectCatalogImageDefinitionBuildClientCancelResponse contains the response from method ProjectCatalogImageDefinitionBuildClient.BeginCancel.
+type ProjectCatalogImageDefinitionBuildClientCancelResponse struct {
+	// placeholder for future response values
+}
+
+// ProjectCatalogImageDefinitionBuildClientGetBuildDetailsResponse contains the response from method ProjectCatalogImageDefinitionBuildClient.GetBuildDetails.
+type ProjectCatalogImageDefinitionBuildClientGetBuildDetailsResponse struct {
+	// Represents a specific build of an Image Definition.
+	ImageDefinitionBuildDetails
+}
+
+// ProjectCatalogImageDefinitionBuildClientGetResponse contains the response from method ProjectCatalogImageDefinitionBuildClient.Get.
+type ProjectCatalogImageDefinitionBuildClientGetResponse struct {
+	// Represents a specific build of an Image Definition.
+	ImageDefinitionBuild
+}
+
+// ProjectCatalogImageDefinitionBuildsClientListByImageDefinitionResponse contains the response from method ProjectCatalogImageDefinitionBuildsClient.NewListByImageDefinitionPager.
+type ProjectCatalogImageDefinitionBuildsClientListByImageDefinitionResponse struct {
+	// Results of the Image Definition Build list operation.
+	ImageDefinitionBuildListResult
+}
+
+// ProjectCatalogImageDefinitionsClientBuildImageResponse contains the response from method ProjectCatalogImageDefinitionsClient.BeginBuildImage.
+type ProjectCatalogImageDefinitionsClientBuildImageResponse struct {
+	// placeholder for future response values
+}
+
+// ProjectCatalogImageDefinitionsClientGetByProjectCatalogResponse contains the response from method ProjectCatalogImageDefinitionsClient.GetByProjectCatalog.
+type ProjectCatalogImageDefinitionsClientGetByProjectCatalogResponse struct {
+	// Represents a definition for an Image.
+	ImageDefinition
+}
+
+// ProjectCatalogImageDefinitionsClientGetErrorDetailsResponse contains the response from method ProjectCatalogImageDefinitionsClient.GetErrorDetails.
+type ProjectCatalogImageDefinitionsClientGetErrorDetailsResponse struct {
+	// List of validator error details. Populated when changes are made to the resource or its dependent resources that impact
+	// the validity of the Catalog resource.
+	CatalogResourceValidationErrorDetails
+}
+
+// ProjectCatalogImageDefinitionsClientListByProjectCatalogResponse contains the response from method ProjectCatalogImageDefinitionsClient.NewListByProjectCatalogPager.
+type ProjectCatalogImageDefinitionsClientListByProjectCatalogResponse struct {
+	// Results of the Image Definition list operation.
+	ImageDefinitionListResult
+}
+
 // ProjectCatalogsClientConnectResponse contains the response from method ProjectCatalogsClient.BeginConnect.
 type ProjectCatalogsClientConnectResponse struct {
 	// placeholder for future response values
@@ -489,6 +579,35 @@ type ProjectEnvironmentTypesClientUpdateResponse struct {
 	ProjectEnvironmentType
 }
 
+// ProjectPoliciesClientCreateOrUpdateResponse contains the response from method ProjectPoliciesClient.BeginCreateOrUpdate.
+type ProjectPoliciesClientCreateOrUpdateResponse struct {
+	// Represents an project policy resource.
+	ProjectPolicy
+}
+
+// ProjectPoliciesClientDeleteResponse contains the response from method ProjectPoliciesClient.BeginDelete.
+type ProjectPoliciesClientDeleteResponse struct {
+	// placeholder for future response values
+}
+
+// ProjectPoliciesClientGetResponse contains the response from method ProjectPoliciesClient.Get.
+type ProjectPoliciesClientGetResponse struct {
+	// Represents an project policy resource.
+	ProjectPolicy
+}
+
+// ProjectPoliciesClientListByDevCenterResponse contains the response from method ProjectPoliciesClient.NewListByDevCenterPager.
+type ProjectPoliciesClientListByDevCenterResponse struct {
+	// Results of the project policy list operation.
+	ProjectPolicyListResult
+}
+
+// ProjectPoliciesClientUpdateResponse contains the response from method ProjectPoliciesClient.BeginUpdate.
+type ProjectPoliciesClientUpdateResponse struct {
+	// Represents an project policy resource.
+	ProjectPolicy
+}
+
 // ProjectsClientCreateOrUpdateResponse contains the response from method ProjectsClient.BeginCreateOrUpdate.
 type ProjectsClientCreateOrUpdateResponse struct {
 	// Represents a project resource.
@@ -498,6 +617,12 @@ type ProjectsClientCreateOrUpdateResponse struct {
 // ProjectsClientDeleteResponse contains the response from method ProjectsClient.BeginDelete.
 type ProjectsClientDeleteResponse struct {
 	// placeholder for future response values
+}
+
+// ProjectsClientGetInheritedSettingsResponse contains the response from method ProjectsClient.GetInheritedSettings.
+type ProjectsClientGetInheritedSettingsResponse struct {
+	// Applicable inherited settings for a project.
+	InheritedSettingsForProject
 }
 
 // ProjectsClientGetResponse contains the response from method ProjectsClient.Get.
@@ -522,6 +647,12 @@ type ProjectsClientListBySubscriptionResponse struct {
 type ProjectsClientUpdateResponse struct {
 	// Represents a project resource.
 	Project
+}
+
+// SKUsClientListByProjectResponse contains the response from method SKUsClient.NewListByProjectPager.
+type SKUsClientListByProjectResponse struct {
+	// Results of the Microsoft.DevCenter SKU list operation.
+	SKUListResult
 }
 
 // SKUsClientListBySubscriptionResponse contains the response from method SKUsClient.NewListBySubscriptionPager.